Refactored and fixed an issue where output paths are incorrectly loaded
Removed Singleton implementation in favour of Dependency Injection.
