+++ MusicBeeNowPlayingOBS.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30320.27
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MusicBeeNowPlayingOBS", "MusicBeeNowPlayingOBS.csproj", "{F5D46BA1-6F21-40EF-9695-46105CCACD08}"
-EndProject
 Project("{911E67C6-3D85-4FCE-B560-20A9C3E3FF48}") = "MusicBee", "..\External References\MusicBeePortable\MusicBee.exe", "{0D3D2D2F-350D-4CF3-989B-E04521CC1DFF}"
 	ProjectSection(DebuggerProjectSystem) = preProject
 		PortSupplier = 00000000-0000-0000-0000-000000000000
@@ -23,6 +21,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MusicBeeExporter", "MusicBeeExporter\MusicBeeExporter.csproj", "{70EC6C50-C2FC-4A89-A864-7E8B62AFE8F9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MusicBeePluginWrapper", "MusicBeePluginWrapper\MusicBeePluginWrapper.csproj", "{F5D46BA1-6F21-40EF-9695-46105CCACD08}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -31,18 +33,26 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F5D46BA1-6F21-40EF-9695-46105CCACD08}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{F5D46BA1-6F21-40EF-9695-46105CCACD08}.Debug|Any CPU.Build.0 = Debug|x86
+		{0D3D2D2F-350D-4CF3-989B-E04521CC1DFF}.Debug|Any CPU.ActiveCfg = Release|x86
+		{0D3D2D2F-350D-4CF3-989B-E04521CC1DFF}.Debug|x86.ActiveCfg = Release|x86
+		{0D3D2D2F-350D-4CF3-989B-E04521CC1DFF}.Release|Any CPU.ActiveCfg = Release|x86
+		{0D3D2D2F-350D-4CF3-989B-E04521CC1DFF}.Release|x86.ActiveCfg = Release|x86
+		{70EC6C50-C2FC-4A89-A864-7E8B62AFE8F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70EC6C50-C2FC-4A89-A864-7E8B62AFE8F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70EC6C50-C2FC-4A89-A864-7E8B62AFE8F9}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{70EC6C50-C2FC-4A89-A864-7E8B62AFE8F9}.Debug|x86.Build.0 = Debug|Any CPU
+		{70EC6C50-C2FC-4A89-A864-7E8B62AFE8F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70EC6C50-C2FC-4A89-A864-7E8B62AFE8F9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{70EC6C50-C2FC-4A89-A864-7E8B62AFE8F9}.Release|x86.ActiveCfg = Release|Any CPU
+		{70EC6C50-C2FC-4A89-A864-7E8B62AFE8F9}.Release|x86.Build.0 = Release|Any CPU
+		{F5D46BA1-6F21-40EF-9695-46105CCACD08}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F5D46BA1-6F21-40EF-9695-46105CCACD08}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F5D46BA1-6F21-40EF-9695-46105CCACD08}.Debug|x86.ActiveCfg = Debug|x86
 		{F5D46BA1-6F21-40EF-9695-46105CCACD08}.Debug|x86.Build.0 = Debug|x86
 		{F5D46BA1-6F21-40EF-9695-46105CCACD08}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F5D46BA1-6F21-40EF-9695-46105CCACD08}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F5D46BA1-6F21-40EF-9695-46105CCACD08}.Release|x86.ActiveCfg = Release|x86
 		{F5D46BA1-6F21-40EF-9695-46105CCACD08}.Release|x86.Build.0 = Release|x86
-		{0D3D2D2F-350D-4CF3-989B-E04521CC1DFF}.Debug|Any CPU.ActiveCfg = Release|x86
-		{0D3D2D2F-350D-4CF3-989B-E04521CC1DFF}.Debug|x86.ActiveCfg = Release|x86
-		{0D3D2D2F-350D-4CF3-989B-E04521CC1DFF}.Release|Any CPU.ActiveCfg = Release|x86
-		{0D3D2D2F-350D-4CF3-989B-E04521CC1DFF}.Release|x86.ActiveCfg = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
