+++ MusicBeePluginWrapper/Configurations/SettingsForm.resx
