+++ MusicBeePluginWrapper/MusicBeePluginWrapper.csproj
@@ -9,7 +9,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>MusicBeePlugin</RootNamespace>
-    <AssemblyName>MB_NowPlayingOBS</AssemblyName>
+    <AssemblyName>MB_MusicBeeExporter</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <FileUpgradeFlags>
@@ -72,7 +72,8 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="Newtonsoft.Json, Version=12.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
-      <HintPath>packages\Newtonsoft.Json.12.0.3\lib\net40\Newtonsoft.Json.dll</HintPath>
+      <HintPath>..\packages\Newtonsoft.Json.12.0.3\lib\net40\Newtonsoft.Json.dll</HintPath>
+      <EmbedInteropTypes>False</EmbedInteropTypes>
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Data" />
@@ -81,14 +82,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ExporterConfiguration\CustomSettings.cs" />
     <Compile Include="MusicBeeInterface.cs" />
-    <Compile Include="NowPlayingOBSExporter.cs" />
+    <Compile Include="MusicBeeExporterInterface.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ExporterConfiguration\SettingsForm.cs">
+    <Compile Include="Configurations\SettingsForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="ExporterConfiguration\SettingsForm.Designer.cs">
+    <Compile Include="Configurations\SettingsForm.Designer.cs">
       <DependentUpon>SettingsForm.cs</DependentUpon>
     </Compile>
   </ItemGroup>
@@ -119,11 +119,17 @@
     </BootstrapperPackage>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="ExporterConfiguration\SettingsForm.resx">
+    <EmbeddedResource Include="Configurations\SettingsForm.resx">
       <DependentUpon>SettingsForm.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <ProjectReference Include="..\MusicBeeExporter\MusicBeeExporter.csproj">
+      <Project>{70ec6c50-c2fc-4a89-a864-7e8b62afe8f9}</Project>
+      <Name>MusicBeeExporter</Name>
+      <Private>True</Private>
+    </ProjectReference>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PreBuildEvent>
