+++ MusicBeePluginWrapper/Properties/AssemblyInfo.cs
