+++ MusicBeePluginWrapper/app.config
@@ -1,0 +1,7 @@
+<?xml version="1.0"?>
+
+<configuration>
+  <startup>
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.0,Profile=Client" />
+  </startup>
+</configuration>
