+++ MusicBeePluginWrapper/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
+
 <packages>
   <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net40-client" />
 </packages>
